Rename paths to project_paths and add utilities
Rename core/paths.py to core/project_paths.py and update imports across the package and tests. Introduce DLCPathContext dataclass and a find_nearest_config helper, and refactor find_nearest_project_root to reuse that helper. Update all modules/tests that referenced the old paths module (_reader, _widgets, core/io, core/metadata, core/remap, misc, and tests) to import from project_paths.
